Implementacion de estructura de MVVM, Implemenatacion de Fragmenst para uan mejor optimizacion de proyecto
