+++ app/src/main/java/com/sigmas/dogapp/Network/RetroFitRazas.kt
@@ -1,0 +1,15 @@
+package com.sigmas.dogapp.Network
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object RetrofitRazas {
+    private const val BASE_URL = "https://dog.ceo/api/"
+
+    val instance: Retrofit by lazy {
+        Retrofit.Builder()
+            .baseUrl(BASE_URL)
+            .addConverterFactory(GsonConverterFactory.create())
+            .build()
+    }
+}
