+++ app/src/main/java/com/sigmas/dogapp/Ui/Fragment/EditarCitaFragment.kt
@@ -1,0 +1,27 @@
+package com.sigmas.dogapp.ui.fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.sigmas.dogapp.databinding.FragmentEditarCitaBinding
+
+class EditarCitaFragment : Fragment() {
+
+    private var _binding: FragmentEditarCitaBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentEditarCitaBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
