+++ app/src/main/java/com/sigmas/dogapp/Ui/Fragment/NuevaCitaFragment.kt
@@ -1,0 +1,27 @@
+package com.sigmas.dogapp.Ui.Fragment
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.sigmas.dogapp.databinding.ActivityNewAppointmentBinding
+
+class NuevaCitaFragment : Fragment() {
+
+    private var _binding: ActivityNewAppointmentBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = ActivityNewAppointmentBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
