+++ app/src/main/java/com/sigmas/dogapp/Ui/MainActivity.kt
@@ -1,0 +1,12 @@
+package com.example.dogapp.ui
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.sigmas.dogapp.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
