+++ app/src/main/java/com/sigmas/dogapp/ViewModel/CitaDetailViewModelFactory.kt
@@ -1,0 +1,14 @@
+package com.sigmas.dogapp.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.sigmas.dogapp.Repository.CitaRepository
+
+class CitaDetailViewModelFactory(private val repository: CitaRepository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(CitaDetailViewModel::class.java)) {
+            return CitaDetailViewModel(repository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
