+++ app/src/main/java/com/sigmas/dogapp/ViewModel/HomeViewModelFactory.kt
@@ -1,0 +1,14 @@
+package com.sigmas.dogapp.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.sigmas.dogapp.Repository.CitaRepository
+
+class HomeViewModelFactory(private val repository: CitaRepository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(HomeViewModel::class.java)) {
+            return HomeViewModel(repository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
