+++ app/src/main/java/com/sigmas/dogapp/ViewModel/NewAppointmentViewModelFactory.kt
@@ -1,0 +1,15 @@
+package com.sigmas.dogapp.viewmodel
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import com.sigmas.dogapp.Repository.CitaRepository
+import com.sigmas.dogapp.ViewModel.NewAppointmentViewModel
+
+class NewAppointmentViewModelFactory(private val repository: CitaRepository) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>): T {
+        if (modelClass.isAssignableFrom(NewAppointmentViewModel::class.java)) {
+            return NewAppointmentViewModel(repository) as T
+        }
+        throw IllegalArgumentException("Unknown ViewModel class")
+    }
+}
