--- app/src/main/java/com/sigmas/dogapp/view/Data/AppDatabase.kt
@@ -1,30 +1,0 @@
-package com.sigmas.dogapp.view.Data
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import com.sigmas.dogapp.view.Data.Model.Cita
-
-@Database(entities = [Cita::class], version = 1, exportSchema = false)
-abstract class AppDatabase : RoomDatabase() {
-
-    abstract fun citaDao(): CitaDao
-
-    companion object {
-        @Volatile
-        private var INSTANCE: AppDatabase? = null
-
-        fun getDatabase(context: Context): AppDatabase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    AppDatabase::class.java,
-                    "dogapp_database"
-                ).build()
-                INSTANCE = instance
-                instance
-            }
-        }
-    }
-}
