--- app/src/main/java/com/sigmas/dogapp/view/Data/Model/RazasResponse.kt
@@ -1,7 +1,0 @@
-package com.sigmas.dogapp.view.Data.Model
-
-// [Modelo de respuesta de razas de perros] (Representa la estructura de respuesta del API dog.ceo para obtener la lista de razas)
-data class RazasResponse(
-    val message: Map<String, List<String>>, // [Listado de razas y subrazas] (Clave: raza principal, Valor: lista de subrazas)
-    val status: String                      // [Estado de la respuesta] ("success" o "error" según el resultado)
-)
