--- app/src/main/res/layout/activity_home.xml
@@ -1,90 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?> <!-- [Pantalla Principal: Administrador de Citas] (Muestra la lista de citas y permite añadir nuevas) -->
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@android:color/white">
-
-    <!-- [Encabezado Curvado] (Header con fondo gris oscuro curvado en esquina inferior derecha) -->
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/headerLayout"
-        android:layout_width="0dp"
-        android:layout_height="200dp"
-        android:background="@drawable/rounded_bottom_right"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent">
-
-        <!-- [Título Principal] (Texto blanco grande con salto de línea indicando sección "Administrador de Citas") -->
-        <TextView
-            android:id="@+id/tvAdministrador"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="30dp"
-            android:text="Administrador\n de Citas"
-            android:textColor="@android:color/white"
-            android:textSize="30sp"
-            android:textStyle="bold"
-            android:gravity="center_horizontal"
-            android:textAlignment="center"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <!-- [Logo del Perro] (Imagen decorativa de un perro en la parte superior derecha del header) -->
-        <ImageView
-            android:id="@+id/imgDog"
-            android:layout_width="100dp"
-            android:layout_height="90dp"
-            android:layout_marginTop="50dp"
-            android:layout_marginEnd="18dp"
-            android:scaleType="centerCrop"
-            android:src="@drawable/lg_dog"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <!-- [Texto DogApp] (Etiqueta decorativa debajo del logo del perro) -->
-        <TextView
-            android:id="@+id/tvDogApp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="15dp"
-            android:layout_marginEnd="15dp"
-            android:text="DogApp"
-            android:textColor="@android:color/white"
-            android:textSize="20sp"
-            android:textStyle="bold"
-            app:layout_constraintEnd_toEndOf="@id/imgDog"
-            app:layout_constraintTop_toBottomOf="@id/imgDog" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <!-- [Listado de Citas] (RecyclerView que muestra todas las citas agendadas por el usuario) -->
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerViewCitas"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        android:contentDescription="Lista de citas de mascotas"
-        android:padding="16dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/headerLayout"
-        app:layout_constraintVertical_bias="0.0" />
-
-    <!-- [Botón Flotante de Añadir Cita] (Botón redondo flotante en esquina inferior derecha para crear nueva cita) -->
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fabAddCita"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="20dp"
-        android:layout_marginBottom="60dp"
-        android:backgroundTint="#F05861"
-        android:src="@drawable/ic_add"
-        app:tint="@android:color/white"
-        android:contentDescription="Agregar nueva cita"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"/>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
