+++ app/src/main/res/layout/activity_login.xml
@@ -50,4 +50,17 @@
         android:layout_marginBottom="70dp"
         app:layout_constraintVertical_bias="0.5"/>
 
+    <Button
+        android:id="@+id/btnAcceder"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Acceder sin huella"
+        android:textColor="@android:color/white"
+        android:backgroundTint="#D23B69"
+        android:layout_marginBottom="20dp"
+        app:layout_constraintTop_toBottomOf="@+id/fingerprintAnimation"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"/>
+
 </androidx.constraintlayout.widget.ConstraintLayout>
