+++ app/src/main/res/layout/fragment_cita_detail.xml
