+++ build.gradle.kts
@@ -1,6 +1,9 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
+// build.gradle.kts (Project-level)
+
 plugins {
     alias(libs.plugins.android.application) apply false
     alias(libs.plugins.kotlin.android) apply false
     alias(libs.plugins.kotlin.compose) apply false
+    id("androidx.navigation.safeargs") version "2.7.7" apply false
 }
+
